+++ CHANGELOG.md
@@ -9,11 +9,17 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ## Unreleased
 
-## [26.4.7] - 2021-06-29
+## [26.4.8] - 2021-06-29
 
 ### Fixed
 
-- /me route forwarding
+- me route forwarding
+
+## [26.4.7] - 2021-06-22
+
+### Added
+
+- SC-9148 - Add migration for change of school year on BRB
 
 ## [26.4.6] - 2021-06-24